Warning fixes

+++ src/prodbg/core/Alloc.cpp
@@ -12,7 +12,7 @@ namespace prodbg
 void* alloc_zero(int size)
 {
 	void* t = malloc((size_t)size);
-	memset(t, 0, size);
+	memset(t, 0, (size_t)size);
 	return t;
 }
